JSON Streaming and Listening Server
BIG UPDATE: Now the scraping script checks for updates every 5 seconds
and makes sure any new changes are actually new. It keeps a list of the
links it has been too so it doesn’t stream redundant entries. It also
is now formatted to JSON. I’ll probably have to make some changes to
this object up for now I am  super happy. We are at the stage where we
can start playing with max.